updates 4.28
new ideas was scrapped to change the name to page1 so that I could
attach multiple pages to the code. Did page 2 with a bar across the top
to produce different elements with each selection on the same page.
looked into frames and decided that was a bad idea. also learned about
shadow effects for the buttons and will be playing with making those
prettier. plan on also doing a nav bar for the left to get through the
site and all the pages easier

used same style sheet.