all freaking day
discovered jfiddle, or rather started using it as a reference tool and
code assistance. worked on setting up a menu with a dropdown with the
ultimate goal of being able to click a button to have different
information appear without having to refresh the page. Learned about how
to set up the list to have dropdown menus appear when the links are
hovered over which I may use somewhere else. figured out how to set it
up but want it to be more specific in case I want other lists on the
page (using class or id system to set it apart) didn't get anything done
on the page itself since  I did it all in jfiddle, but all of the notes
are in the new document "menu for page 2" did ask a friend for some help
and he commented that this is the hardest type of list, thanks friend...