page 2 menu looks good
converted javascript to make the buttons do what I want an display a
menu with information on it. cleaned up css and changed the links to
anchor buttons with script. Also came up with some ideas for a menu that
has dynamic buttons that expand when you hover over them. After messing
with it for a while I need to find a way to make the buttons dynamic
without moving everything around on the page but still keep the page
itself responsive. will probably to a gist to work on this and maybe add
it to the main site at a later date. also learned about cursor tag.